move out of utils functions
